+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# 1.0.0 – 2014-06-xx
+
+- Simple implementation of Less import options. Import keyword is applied to all @import statements in task target.
+- Dev: Test with Mocha instead of Nodeunit
+
 # 0.9.1 – 2014-01-18
 
 - Fix documentation for the banner option.